requirements.txt
asgiref==3.2.7
boto3==1.10.23
botocore==1.13.50
certifi==2019.9.11
chardet==3.0.4
colorama==0.4.3
dj-database-url==0.5.0
Django==3.0.4
django-cors-headers==3.2.1
django-filter==2.2.0
django-storages==1.8
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
docutils==0.15.2
gunicorn==20.0.0
idna==2.8
jmespath==0.9.5
Pillow==7.0.0
psycopg2==2.8.4
Pygments==2.6.1
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
s3transfer==0.2.1
six==1.13.0
sqlparse==0.3.1
urllib3==1.25.7
whitenoise==4.1.4
